Fix bugs in Autonomous code
-Fix the recently added functionality for MoveForward to accept negative
input (for going backwards). Spent way too long trying to fix it (turned
out I just had a stray static modifier). Anyway, the backwards
functionality is fully tested and furthermore the functionality for the
sendable chooser to also go backwards in autonomous (this was a desired
functionality after talking to Mr. Cole), since it sets us up for Teleop
better. The new autonomous now is much more modular. Currently, you can
only choose from 3 distances to move forward, and whether to go back the
same distance after it, and spybot, and doing nothing. The Rotate
command is still woefully undertested since I haven't had a drivebase to
be able to test that, I would love it if that could be tested
thoroughly. The spybot command is currently handwritten and likely
doesn't exactly work, although it could. Ideally we would use the
recordable autonomous for the spybot. The third sendable chooser (no
shot vs low goal vs high goal) is currently ONLY used for spybot
command, although that theoretically would be applied to moving forward
as well.
-I also reorganized the SmartDashboard and gave it a fresh touch up and
commit. I also made it so the main image and the Target are both in
C:FRC/PICS/. The images are still stored in the git repository, but if
people could make that directory and copy it there, then we don't need
to deal with the image paths on the dashboard being tied to people's
laptops.
-Any questions while the three of us are gone on Saturday (2/27), go
ahead and shoot me a text, I can probably answer it.