Overhaul the drive of Lenny
1. The winch disables in MoveAimerUp and MoveAimerDown if you're trying
to move it past the endpoint. It reenables when trying to move it in the
right direction
2. Change the setpoint levels to a Levels enum
3. The levels right now are GROUND, START, LOW, and HIGH, where START is
always 0. This will need to be adjusted for new gearboxed, etc.
4. Change name of GyroPID to DeltaPID. was planning on using; never did
5. Removed GoToLevel and replaced with GoToPosition, which can accept a
raw encoder count or a level enum
6. Add joystick buttons L/R Bumper to go straight to GROUND or START,
respectively.
8. GoToPosition drives in velocity mode using a TrajectoryController to
get there
9. There is a timeout in GoToPosition of currently 4 seconds, since we
were having problems at the end of Wednesday.
10. Change use of disableControl() and enableControl() to disable() and
enable() in CurrentSafety, and am using it to disable in GoToPosition