Add climber crap
None of the methods or commands are accessible right now to the gamepad.
They are just abstractly there. To implement climber, set hte
ClimbUp/Down/Stop to the POV, for example, on gamepad2 in OI much like
the winch or tail is set up, and add the LockClimber/ReleaseClimber
commands to buttons. The Climber method encapsulates all the methods and
contains all the constants.

Have fun!