Add safety for winch limit switch
If while Lenny is holding position, the limit switch is inadvertently
triggered, then it will still try to hold its position but that position
will have moved, so it has been observed to slam down because the PID is
so strong to hold position.

The limit switch now only resets the encoder if the encoder is
reasonably close to 0 (currently +/- 150)

This is currently loaded on the robot.